Modeless window feature : ExternalExecutor strikes back, more generic way to communicate with revit from outside of its context
